Automatically start a timer
The host (or anyone) can start a round with “-round”, and the bot will
ping everyone that it’s about to start, and then the next purely numeric
message will start the countdown.

State is tracked via a global for now, which isn’t very pretty, but I
didn’t see any convenient helpers in the discord lib.