Add environment variable to skip audio
The bot has been running for a while with an uncommitted, hard-coded
return in the play_audio() function, because the audio was broken for
some reason that we hadn’t investigated yet. Make that return
configurable so it can be committed – an environment variable seems like
the easiest solution for now.